requirements.txt
python-dotenv==1.0.1
requests==2.32.3
psycopg2-binary==2.9.10
tenacity==9.0.0
transformers==4.45.2
torch>=2.0.0
tqdm==4.66.5
pydantic==2.9.2
fastapi==0.115.5
uvicorn==0.32.1
streamlit==1.40.0
plotly==5.24.1
pandas==2.2.3
